+++ CHANGELOG.md
@@ -1,3 +1,2 @@
 
-
-
+- added [useLeafletIsMoving](../../..#useleafletismoving) & [useLeafletIsZooming](../../..#useleafletiszooming) hooks.